% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_of_onset.R, R/set_vars.R
\name{get_var}
\alias{get_var}
\alias{date_of_onset}
\alias{date_onset}
\alias{id}
\alias{gender}
\alias{date_report}
\alias{age}
\alias{age_group}
\alias{set_vars}
\alias{set_vars<-}
\alias{set_vars.linelist}
\alias{set_vars<-.linelist}
\title{Accessor functfons}
\usage{
get_var(x, var)

date_of_onset(x)

date_onset(x)

id(x)

gender(x)

date_report(x)

age(x)

age_group(x)

set_vars(x, ...)

set_vars(x, name) <- value

\method{set_vars}{linelist}(x, ...)

\method{set_vars}{linelist}(x, name) <- value
}
\arguments{
\item{x}{a linelist object}

\item{var}{the name of a variable as a string}

\item{...}{For \code{set_vars()}, any number of variables defined in
\code{\link[=epivars]{epivars()}} that can be used for mapping or modelling.}

\item{name}{the name of the variable in \code{\link[=epivars]{epivars()}} to assign}

\item{value}{the name of the column in the locations data}
}
\value{
a vector if the epivar can be represented by a single column, a data
frame otherwise (e.g. a location column).
}
\description{
This set of fucntions defines helper functions for accessing pre-defined
epivars in your linelist object.
}
\examples{

toy_data <- messy_data()
cleaned_data <- clean_data(toy_data)
ll <- as_linelist(cleaned_data, 
                  id = "id", 
                  date_onset = "date_of_onset",
                  gender = "gender")
date_of_onset(ll)
id(ll)
gender(ll)

# epivars that haven't been defined for the data set will return an error
try(date_report(ll))
}
