% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{add_epivar}
\alias{add_epivar}
\alias{add_description}
\title{Add or remove elements from the current dictionary}
\usage{
add_epivar(epivar = NULL, hxl = "", description = "")

add_description(epivar = NULL, description = NULL)
}
\arguments{
\item{epivar}{a character describing an epivar or a data frame containing a
dictionary to append to the current dictionary.}

\item{hxl}{a hxl hashtag and metadata}

\item{description}{a character string giving a human-readable description of the
epivar.}
}
\value{
NULL, invisibly
}
\description{
These functions give a finer set of controls to the dictionary, allowing the
user to add rows to the dictionary.
}
\examples{
# preserve current dictionary
od <- get_dictionary()

# add a single row for "date_hospital"
hosp <- data.frame(
  epivar = "date_hospital",
  hxl    = "#date +start",
  description = "date at which patient was hospitalized",
  stringsAsFactors = FALSE
)

add_epivar(hosp)
get_dictionary()

# Use add_description to update descriptions
add_description("date_hospital", "hospitalisation date")
get_dictionary()

# You can export your dictionary for later use
tmp <- tempdir()
rio::export(get_dictionary(), file.path(tmp, "new_dictionary.csv"), format = "|")
head(rio::import(file.path(tmp, "new_dictionary.csv"), format = "|"))
unlink(file.path(tmp, "new_dictionary.csv")) # cleanup

# reset dictionary
set_dictionary(od)
}
