% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.linelist.R
\name{as.data.frame.linelist}
\alias{as.data.frame.linelist}
\title{Convert a linelist object to a data frame}
\usage{
\method{as.data.frame}{linelist}(x, ...)
}
\arguments{
\item{x}{a linelist object}

\item{...}{unused}
}
\value{
a data frame without the associated metadata
}
\description{
Convert a linelist object to a data frame
}
\examples{
md <- messy_data()
cd <- clean_data(md)
ll <- as_linelist(cd, 
                  id = "id", 
                  date_onset = "date_of_onset", 
                  gender = "gender",
                  geo = c("lon", "lat")
                 )
identical(cd, as.data.frame(ll))
}
\author{
Zhian N. Kamvar
}
