% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_spelling.R
\name{clean_spelling}
\alias{clean_spelling}
\title{Rename values in a vector based on a dictionary}
\usage{
clean_spelling(x = character(), dictionary = data.frame())
}
\arguments{
\item{x}{a character vector}

\item{dictionary}{a two-column matrix or data frame defining mis-spelled words in
the first column and replacements in the second column}
}
\description{
Rename values in a vector based on a dictionary
}
\examples{

corrections <- data.frame(
  bad = c("foubar", "foobr", "fubar", NA, "unknown"),
  good = c("foobar", "foobar", "foobar", "missing", "missing"),
  stringsAsFactors = FALSE
)
head(corrections)
my_data <- c(letters[1:5], sample(corrections$bad, 10, replace = TRUE))
clean_spelling(my_data, corrections)

# The function will give you a warning if the dictionary does not
# match the data
clean_spelling(letters, corrections)

}
\author{
Zhian N. Kamvar
}
