% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_variable_names.R
\name{clean_variable_names}
\alias{clean_variable_names}
\title{Clean variable names}
\usage{
clean_variable_names(x, ...)
}
\arguments{
\item{x}{a \code{data.frame}}

\item{...}{further arguments passed to \code{epitrix::clean_labels}; the most
important is \code{sep}, which refers to the separator used between words,
and defaults to the underscore \code{_}.}
}
\description{
This function takes standardises the variable names in a
\code{data.frame}. It uses the standardisation implemented by
\code{\link[epitrix]{clean_labels}} in the \code{epitrix} package. See
\code{?epitrix::clean_labels} for more information.
}
\author{
Thibaut Jombart
}
