% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Internal function to collapse text}
\usage{
collapse(x, sep = ", ")
}
\arguments{
\item{x}{a \code{character} vector}

\item{sep}{a separator to be used to collate the text}
}
\description{
Internal function to collapse text
}
