% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_data.R
\name{compare_classes}
\alias{compare_classes}
\title{Compare vectors of classes}
\usage{
compare_classes(ref, x)
}
\description{
The function returns \code{NULL} if the classes are the same, and a named list of
character strings if there are differences.
}
