% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_data.R
\name{compare_data}
\alias{compare_data}
\alias{compare_data.default}
\alias{compare_data.data.frame}
\alias{print.data_comparison}
\title{Compare structures of two datasets}
\usage{
compare_data(ref, x, ...)

\method{compare_data}{default}(ref, x, ...)

\method{compare_data}{data.frame}(ref, x, ...)

\method{print}{data_comparison}(x, ...)
}
\arguments{
\item{ref}{the reference \code{data.frame}}

\item{x}{a \code{data.frame} to be compared against a reference}
}
\description{
This function extracts the structures of two \code{data.frames} and compares them,
issuing a series of diagnostics.
}
\details{
The comparison relies on checking differences in:
\itemize{
\item names of columns
\item classes of the columns (only the first class is used)
\item values of the categorical variables
}
}
\examples{

## no differences
compare_data(iris, iris)

## different dimensions
compare_data(iris, iris[-1, -2])
compare_data(iris[-1, -2], iris) # inverse

## one variable in common but different class and content
compare_data(iris,
             data.frame(Species = letters,
                        stringsAsFactors = FALSE))
}
\author{
Thibaut Jombart
}
