% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_data.R
\name{compare_values}
\alias{compare_values}
\title{Compare values of categorical variables}
\usage{
compare_values(ref, x)
}
\description{
The function returns \code{NULL} if the categories of categorical variables are
the same, and a named list of character strings if there are differences.
}
