% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_labels}
\alias{default_labels}
\title{Default labels for a linelist}
\usage{
default_labels(...)
}
\arguments{
\item{...}{A series of labels provided as \verb{"label1", "label2"} to add
labels or optionally \code{"id" = "new_id"} to rename defaults}
}
\value{
vector of labels
}
\description{
Default labels for a linelist
}
