% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_types}
\alias{default_types}
\title{Get the default types for linelist variables}
\usage{
default_types(...)
}
\arguments{
\item{...}{a series of vectors containing types to append to the defaults.}
}
\value{
list of types
}
\description{
Get the default types for linelist variables
}
\note{
Use the helper function \code{type()} to generate known collections of types.
}
