% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_vars}
\alias{default_vars}
\title{Default column names for a linelist}
\usage{
default_vars(...)
}
\arguments{
\item{...}{A series of column names provided as \verb{col1, col2} to add columns
or optionally \verb{id = 'ID', date_onset = 'date_symptom_onset'} to rename
defaults}
}
\value{
vector of variables
}
\description{
Default column names for a linelist
}
