% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivars.R
\name{epivars}
\alias{epivars}
\alias{linelist.epivars}
\title{Linelist Global Variables}
\usage{
epivars(..., set = FALSE, reset = FALSE)
}
\arguments{
\item{...}{quoted varaibles to add to the default variables}

\item{set}{when \code{TRUE}, the variables provided in \code{...} will be added to the
global variables. Defaults to `FALSE``}

\item{reset}{when \code{TRUE}, the global variables are reset to the default
variables listed above. Defaults to \code{FALSE}}
}
\description{
This defines the legal variables that linelist will recognise.
}
\examples{

# see the default varaibles
epivars()

# Equivalent
getOption("linelist.epivars")

# If you create a new method and need other varaibles, or just want a shorter
# representation, they can be added to your options:

epivars("random", set = TRUE)

# You can also reset the variables
epivars(reset = TRUE)
}
