% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_categories.R
\name{get_categories}
\alias{get_categories}
\title{List unique values of categorical data}
\usage{
get_categories(x)
}
\description{
This function is for internal use only. It extracts and sorts levels of factors, and unique values of characters.
}
