% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Retrieve metadata from a linelist object}
\usage{
get_meta(dat, dictionary = TRUE)
}
\arguments{
\item{dat}{a \code{linelist} object}

\item{dictionary}{if \code{TRUE}, a column called "\emph{epivar}" is appended to the
metadata table, indicating which standard epi variable columns correspond
to; of \code{FALSE}, the class of the column replaces it}
}
\value{
a \code{data.frame} representing metatdata
}
\description{
Retrieve metadata from a linelist object
}
\examples{
dat <- clean_data(messy_data())
ll  <- as_linelist(dat,
                  id = "id", 
                  date_onset = "date_of_onset", 
                  gender = "gender",
                  geo = c("lon", "lat")
                 )
get_meta(ll)
get_meta(ll, dict = FALSE)
}
