% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structure.R
\name{get_structure}
\alias{get_structure}
\title{Extract the structure of a dataset}
\usage{
get_structure(x)
}
\arguments{
\item{x}{a \code{data.frame} to be checked}
}
\value{
A list containing:
\itemize{
\item \code{dim}: the dimensions of the dataset (rows, columns)
\item \code{names}: the names of the dataset
\item \code{classes}: the corresponding classes (if a column has several classes, only
the first one is kept)
\item \code{values}: the list of sorted, unique values for each categorical column
(\code{factor} or \code{character})
}
}
\description{
This function extracts the structure and content of a \code{data.frame}, storing
the following information:
}
\details{
\itemize{
\item names and order of columns
\item class of the columns
\item values of the columns, for \code{factors} and \code{characters}
}
}
\examples{

head(iris)
get_structure(iris)
}
\author{
Thibaut Jombart
}
