% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/please-change.R
\name{hello_recon}
\alias{hello_recon}
\title{Hello World}
\usage{
hello_recon(input)
}
\arguments{
\item{input}{you should always document the paramters.
Including the expected data type.}
}
\description{
This is an example of how to create and document exported functions.
}
