% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_find_classes.R
\name{i_find_classes}
\alias{i_find_classes}
\title{Find classes of the columns of a data.frame}
\usage{
i_find_classes(x)
}
\description{
Internal function. Returns a vector of characters indicating the classes of
the different variables of a data.frame. In the case of multiple classes, we
only return the first one.
}
\author{
Thibaut Jombart
}
