% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_find_POSIX.R
\name{i_find_POSIX}
\alias{i_find_POSIX}
\title{Find the columns of a data.frame which are POSIXct or POSIXlt}
\usage{
i_find_POSIX(x)
}
\description{
Internal function. Returns a vector of indices of columns of a data.frame
which are POSIXct or POSIXlt.
}
\author{
Thibaut Jombart
}
