% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_defaults.R
\name{label_defaults}
\alias{label_defaults}
\title{Default labels recognized by \pkg{linelist}}
\usage{
label_defaults()
}
\value{
A character vector with the default labels recognized by
\pkg{linelist}
}
\description{
Default labels recognized by \pkg{linelist}
}
\examples{
label_defaults()

}
