% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linelist_example.R
\name{linelist_example}
\alias{linelist_example}
\title{show the path to a linelist example file}
\usage{
linelist_example(name = NULL)
}
\arguments{
\item{name}{the name of a linelist example file}
}
\value{
a path to a linelist example file
}
\description{
show the path to a linelist example file
}
\examples{
linelist_example() # list all of the example files

# read in example spelling dictionary
sd <- linelist_example("spelling-dictionary.csv")
read.csv(sd, stringsAsFactors = FALSE)

# read in example coded data
coded_data <- linelist_example("coded-data.csv")
coded_data <- read.csv(coded_data, stringsAsFactors = FALSE)
str(coded_data)
coded_data$date <- as.Date(coded_data$date)
}
\author{
Zhian N. Kamvar
}
