% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_epivars.R
\name{list_epivars}
\alias{list_epivars}
\title{List registred epivars in a linelist object}
\usage{
list_epivars(x, simple = FALSE, all_dictionary = FALSE)
}
\arguments{
\item{x}{a \code{linelist} object}

\item{simple}{a logical indicating if a single vector of names of present
\code{epivars} should be returned (\code{TRUE}); otherwise the output will contain
more information as a \code{data.frame} (\code{FALSE}, default)}

\item{all_dictionary}{if \code{FALSE} (default), only the \emph{epivars} present in the
object are documented; otherwise, all the available dictionary is appended
to the output}
}
\value{
if \code{simple} is \code{TRUE}, a vector of characters giving the names of
available \code{epivariables}; otherwise, a \code{data.frame} giving more information
about these variables
}
\description{
List registred epivars in a linelist object
}
\examples{
dat <- clean_data(messy_data())
ll  <- as_linelist(dat,
                  id = "id", 
                  date_onset = "date_of_onset", 
                  gender = "gender",
                  geo = c("lon", "lat")
                 )
list_epivars(ll)
list_epivars(ll, simple = TRUE)
list_epivars(ll, all_dictionary = TRUE)
}
