% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup}
\alias{lookup}
\title{Lookup a column name defined by an epivar}
\usage{
lookup(x, epivar = NULL)
}
\arguments{
\item{x}{a \link[=as_linelist]{linelist} object.}

\item{epivar}{the name of an epivar defined in the \link[=get_dictionary]{dictionary}.}
}
\value{
a character vector of the column name
}
\description{
For use in dplyr pipelines and other functions that need to know the names
of columns, the lookup function looks up the column name for a given epivar
in your data set.
}
\examples{
dat <- clean_data(messy_data())
ll  <- as_linelist(dat,
                   id = "id", 
                   gender = "gender",
                   date_onset = "date_of_onset", 
                   geo = c("lon", "lat")
                  )

}
