% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_epivars.R
\name{mask}
\alias{mask}
\alias{mask.default}
\alias{mask.linelist}
\alias{unmask}
\alias{unmask.default}
\alias{unmask.linelist}
\alias{unmask.data.frame}
\title{Reversibly rename columns in your linelist with epivars}
\usage{
mask(x)

\method{mask}{default}(x)

\method{mask}{linelist}(x)

unmask(x)

\method{unmask}{default}(x)

\method{unmask}{linelist}(x)

\method{unmask}{data.frame}(x)
}
\arguments{
\item{x}{a linelist object}
}
\value{
\itemize{
\item \code{mask()} a linelist object with columns renamed by their
corresponding epivars and an attribute called "masked-linelist" containing
a vector of the original column names named by their corresponding epivars
\item \code{unmask()} a linelist object
}
}
\description{
Reversibly rename columns in your linelist with epivars
}
\examples{
md <- messy_data(10)
cd <- clean_data(md)
ll <- as_linelist(cd, 
                  id = "id", 
                  date_onset = "date_of_onset", 
                  gender = "gender",
                  case_definition = "epi_case_definition",
                  geo_lon = "lon",
                  geo_lat = "lat"
                 )
ll

# masking epivars changes the column names
print(llm <- mask(ll))

# unmasking reverses the operation
unmask(llm)

# you can safely drop columns
unmask(llm[grepl("gender|geo", names(llm))])
}
\author{
Zhian N. Kamvar
}
