% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_epivars.R
\name{mask_epivars}
\alias{mask_epivars}
\alias{unmask_epivars}
\title{Reversibly rename columns in your data frame with epivars}
\usage{
mask_epivars(x)

unmask_epivars(x)
}
\arguments{
\item{x}{a linelist object}
}
\value{
\itemize{
\item \code{mask_epivars()} a linelist object with columns renamed by their
corresponding epivars and an attribute called "masked-linelist" containing
a vector of the original column names named by their corresponding epivars
\item \code{unmask_epivars()} a linelist object
}
}
\description{
Reversibly rename columns in your data frame with epivars
}
\examples{
md <- messy_data(10)
cd <- clean_data(md)
ll <- as_linelist(cd, 
                  id = "id", 
                  date_onset = "date_of_onset", 
                  gender = "gender",
                  case_definition = "epi_case_definition",
                  geo_lon = "lon",
                  geo_lat = "lat"
                 )
ll

# masking epivars changes the column names
print(llm <- mask_epivars(ll))

# unmasking reverses the operation
unmask_epivars(llm)

# you can safely drop columns
unmask_epivars(llm[grepl("gender|geo", names(llm))])
}
\author{
Zhian N. Kamvar
}
