% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messy_data.R
\name{messy_data}
\alias{messy_data}
\title{Function for generating messy data}
\usage{
messy_data(n = 20)
}
\arguments{
\item{n}{the number of cases to generate}
}
\value{
a data frame with "messy" data
}
\description{
Function for generating messy data
}
\examples{
messy_data(n = 5)
messy_data()
}
