% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_defaults.R
\name{modify_defaults}
\alias{modify_defaults}
\title{Modify defaults in a list}
\usage{
modify_defaults(defaults, x, strict = TRUE)
}
\arguments{
\item{defaults}{a named list of default values}

\item{x}{a named list of modified values, to replace the defaults}

\item{strict}{a \code{logical} indicating if allowed modification should be
restricted to variables existing in the defaults (\code{TRUE}); if \code{FALSE}, new
variables will allowed and added to the output}
}
\value{
A named \code{list}.
}
\description{
Internal function. This is used to modify default values provided in a named
list.
}
\author{
Code by Rich FitzJohn borrowed from outbreaker2
}
