% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{names<-.linelist}
\alias{names<-.linelist}
\title{Rename columns of a linelist}
\usage{
\method{names}{linelist}(x) <- value
}
\arguments{
\item{x}{a \code{linelist} object}

\item{value}{a \code{character} vector to set the new names of the columns of \code{x}}
}
\value{
a \code{linelist} with new column names
}
\description{
This function can be used to rename the columns a \code{linelist}, adjusting tags
as needed.
}
\seealso{
\url{rename.linelist} for \code{dplyr}-style renaming of
columns
}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
