% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.linelist.R
\name{print.linelist}
\alias{print.linelist}
\title{Function for printing linelist objects}
\usage{
\method{print}{linelist}(x, ..., show_epivars = FALSE)
}
\arguments{
\item{x}{a \code{linelist} object}

\item{...}{further arguments passed to \code{print.data.frame}}

\item{show_epivars}{a logical indicating whether \code{epivars} should be shown
after the \code{data.frame} (\code{TRUE}), or not (\code{FALSE}, default); note that
\code{epivars} are only shown if present}
}
\description{
The \code{print} method of \code{linelist} objects shows \emph{epivars} below the names of
the variables.
}
\examples{

x <- messy_data(5)
x <- tibble::as_tibble(clean_data(x))
x <- as_linelist(x, 
                 id = "id", 
                 gender = "gender",
                 geo_lon = "lon",
                 geo_lat = "lat"
                )
x
print(x, show_epivars = TRUE)
}
