% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_tags.R
\name{prune_tags}
\alias{prune_tags}
\title{Prune tags after changing columns of a linelist}
\usage{
prune_tags(x, warn = TRUE)
}
\arguments{
\item{x}{\code{linelist} object}

\item{warn}{a \code{logical} indicating if a warning should be issued if some tagged
variables have been lost; defaults to \code{TRUE}}
}
\value{
The function returns a \code{linelist} object.
}
\description{
Internal function. Used to remove tags whose variable has been removed after
subsetting the columns of a \code{linelist} object.
}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
