% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_label.R, R/labels.R, R/labels_df.R,
%   R/lost_labels_action.R, R/set_labels.R, R/type.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{has_label}
\alias{labels}
\alias{labels_df}
\alias{lost_labels_action}
\alias{get_lost_labels_action}
\alias{set_labels}
\alias{type}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{safeframe}{\code{\link[safeframe:lost_labels_action]{get_lost_labels_action}}, \code{\link[safeframe]{has_label}}, \code{\link[safeframe]{labels}}, \code{\link[safeframe]{labels_df}}, \code{\link[safeframe]{lost_labels_action}}, \code{\link[safeframe]{set_labels}}, \code{\link[safeframe]{type}}}
}}

