% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_tags.R
\name{restore_tags}
\alias{restore_tags}
\title{Restore tags of a linelist}
\usage{
restore_tags(x, tags, lost_action = c("error", "warning", "none"))
}
\arguments{
\item{x}{a \code{data.frame}}

\item{tags}{a list of tags as returned by \code{tags()}; if default values are
missing, they will be added to the new list of tags}

\item{lost_action}{a \code{character} indicating the behaviour to adopt when tagged
variables have been lost: "error" (default) will issue an error; "warning"
will issue a warning; "none" will do nothing}
}
\value{
The function returns a \code{linelist} object with updated tags.
}
\description{
This function is used to restore tags of a \code{linelist} object which may have
lost its tags after handling e.g. through \code{dplyr} verbs. Specific actions can
be triggered when some of the tagged variables have disappeared from the
object.
}
\seealso{
\url{prune_tags} for removing tags which have lost their
variables
}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
