% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_epivars.R
\name{set_epivars}
\alias{set_epivars}
\alias{reset_epivars}
\alias{set_epivars.linelist}
\alias{reset_epivars.linelist}
\alias{set_epivars.data.frame}
\alias{reset_epivars.data.frame}
\title{Add epivars to a linelist object}
\usage{
set_epivars(x, ...)

reset_epivars(x, ...)

\method{set_epivars}{linelist}(x, ...)

\method{reset_epivars}{linelist}(x, ...)

\method{set_epivars}{data.frame}(x, ...)

\method{reset_epivars}{data.frame}(x, ...)
}
\arguments{
\item{x}{a \code{linelist} object}

\item{...}{a series of column names named after epivars}
}
\value{
a \code{linelist} object with new epivars ammended
}
\description{
These functions can be used to change the \code{epivars} registred in a \code{linelist}
object.
}
\details{
\itemize{
\item \code{set_epivars} adds new \code{epivars} to the existing ones.
\item \code{reset_epivars}  replaces the existing \code{epivars} with new ones.
}
}
\examples{
# define some toy data
cd <- clean_data(messy_data())

# Use as_linelist to define your epivars ---------------------
ll <- as_linelist(cd,
  "gender"          = "gender",
  "id"              = "id",
  "geo"             = c("lon", "lat"),
  "case_definition" = "epi_case_definition"
)
list_epivars(ll)

# add a new epivar -------------------------------------------
ll <- set_epivars(ll, "date_onset" = "date_of_onset")
list_epivars(ll)

# change an epivar column ------------------------------------
# 
# in this case, we want to update the date of onset to reflect the number
# of days from the start 
ll$days_from_start <- as.integer(get_epivars(ll, "date_onset") - as.Date("2018-01-23"))
ll <- set_epivars(ll, "date_onset" = "days_from_start")
list_epivars(ll)

# Remove an epivar label -------------------------------------
ll <- set_epivars(ll, "date_onset" = NULL)
list_epivars(ll)
}
