% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_bracket.R
\name{[.linelist}
\alias{[.linelist}
\title{Subsetting of linelist objets}
\usage{
\method{[}{linelist}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{a \code{linelist} object}

\item{...}{further arguments to be passed to 'print'}
}
\value{
Invisibly returns the object.
}
\description{
This function prints linelist objects
}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
