% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_bracket.R
\name{[.linelist}
\alias{[.linelist}
\alias{[<-.linelist}
\title{Subsetting of linelist objets}
\usage{
\method{[}{linelist}(x, i, j, drop = FALSE, lost_action = "warning")

\method{[}{linelist}(x, i, j, lost_action = "warning") <- value
}
\arguments{
\item{x}{a \code{linelist} object}

\item{i}{a vector of \code{integer} or \code{logical} to subset the rows of the
\code{linelist}}

\item{j}{a vector of \code{character}, \code{integer}, or \code{logical} to subset the
columns of the \code{linelist}}

\item{drop}{a \code{logical} indicating if, when a single column is selected, the
\code{data.frame} class should be dropped to return a simple vector, in which
case the \code{linelist} class is lost as well; defaults to \code{FALSE}}

\item{lost_action}{a \code{character} indicating the behaviour to adopt when some
of the tagged variables are dropped through the subsetting process; can be
"warning" (default), "error", or "none".}
}
\value{
If no drop is happening, a \code{linelist}. Otherwise an atomic vector.
}
\description{
The \verb{[]} operator for \code{linelist} objects behaves like for regular
\code{data.frame} or \code{tibble}, but checks that tagged variables are not lost, and
takes the appropriate action (warning, error, or ignore, depending on the
value of \code{lost_action}) if this is the case.
}
\author{
Thibaut Jombart \url{thibaut@data.org}

Thibaut Jombart \url{thibaut@data.org}
}
