% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_variable.R
\name{tag_variable}
\alias{tag_variable}
\title{Tag a variable using its name}
\usage{
tag_variable(x, var_type, var_name)
}
\arguments{
\item{x}{a \code{data.frame} or a \code{tibble}, with at least one column}

\item{var_type}{a \code{character} indicating the generic data to be tagged}

\item{var_name}{a \code{character} or an \code{integer} indicating the columns of the
dataset corresponding to this type of data}
}
\value{
The function returns the original object with an additional
attribute.
}
\description{
This function will tag a pre-defined type of variable in a \code{data.frame} by
adding a named attribute identifying the column name.
}
\details{
If used several times, the previous tag is removed silently.
}
\examples{
# basic use of the function
template(1)

}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
