% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tags}
\alias{tags}
\title{Get the list of tags in a linelist}
\usage{
tags(x)
}
\arguments{
\item{x}{a \code{linelist} object}
}
\value{
The function returns a named \code{list} where names indicate generic
types of data, and values indicate which column they correspond to.
}
\description{
This function returns the list of tags identifying specific variable types in
a \code{linelist}.
}
\details{
Tags are stored as the \code{tags} attribute of the object.
}
\examples{
# basic use of the function
if (require(outbreaks)) {
measles_hagelloch_1861
x <- make_linelist(measles_hagelloch_1861, date_onset = "date_of_prodrome")
tags(x)
}

}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
