% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_types.R
\name{tags_types}
\alias{tags_types}
\title{List acceptable variable types for tags}
\usage{
tags_types()
}
\value{
A named \code{list}.
}
\description{
This function returns a named list providing the acceptable data types for
the default tags.
}
\examples{
tags_types()
}
\seealso{
\url{tags_defaults} for the default tags
}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
