% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{Title of the function}
\usage{
template(x)
}
\arguments{
\item{x}{...}
}
\value{
The function returns a ...
}
\description{
A short description of what it does, the main inputs and outputs.
}
\details{
Some more information on the function
}
\examples{
# basic use of the function
template(1)

}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
