% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_linelist.R
\name{template_linelist}
\alias{template_linelist}
\title{Generate a template for linelist construction}
\usage{
template_linelist(x = NULL)
}
\arguments{
\item{x}{a data frame of your linelist (optional)}
}
\value{
a printed template that you can use to define the linelist variables
for your data
}
\description{
Generate a template for linelist construction
}
\examples{
dat <- clean_data(messy_data())
template_linelist(dat)
}
