% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_defaults.R
\name{update_defaults}
\alias{update_defaults}
\title{Update default column names for labelling}
\usage{
update_defaults(
  id = "id",
  date_onset = "date_onset",
  date_reporting = "date_reporting",
  date_admission = "date_admission",
  date_discharge = "date_discharge",
  date_outcome = "date_outcome",
  date_death = "date_death",
  gender = "gender",
  age = "age",
  location = "location",
  occupation = "occupation",
  hcw = "hcw",
  outcome = "outcome"
)
}
\arguments{
\item{id}{default column name to replace, using \code{id = "new_id"}}

\item{date_onset}{default column name to replace, using
\code{date_onset = "new_date_onset"}}

\item{date_reporting}{default column name to replace, using
\code{date_reporting = "new_date_reporting"}}

\item{date_admission}{default column name to replace, using
\code{date_admission = "new_date_admission"}}

\item{date_discharge}{default column name to replace, using
\code{date_discharge = "new_date_discharge"}}

\item{date_outcome}{default column name to replace, using
\code{date_outcome = "new_date_outcome"}}

\item{date_death}{default column name to replace, using
\code{date_death = "new_date_death"}}

\item{gender}{default column name to replace, using \code{gender = "new_gender"}}

\item{age}{default column name to replace, using \code{age = "new_age"}}

\item{location}{default column name to replace, using
\code{location = "new_location"}}

\item{occupation}{default column name to replace, using
\code{occupation = "new_occupation"}}

\item{hcw}{default column name to replace, using \code{hcw = "new_hcw"}}

\item{outcome}{default column name to replace, using
\code{outcome = "new_outcome"}}
}
\value{
A named \code{list}.
}
\description{
A function to update default column names for labelling.
}
\details{
You can use this function to update one, all, or any variation of column
names. If run without arguments, it returns the same as \code{label_defaults()}.
}
