% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{update_or_append}
\alias{update_or_append}
\title{Internal. Update or append to a vector}
\usage{
update_or_append(x, ...)
}
\arguments{
\item{x}{origin vector}

\item{...}{(named) list of values to update or append. \code{hcw = "hacw"} will
update the origin vector if \code{hcw} exists, otherwise append it. If there is
no expected equivalent in the origin, write \verb{"newVar1", "newVar2"}}
}
\value{
vector
}
\description{
Internal. Update or append to a vector
}
