% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_tags.R
\name{validate_tags}
\alias{validate_tags}
\title{Checks the tags of a linelist object}
\usage{
validate_tags(x, allow_extra = FALSE)
}
\arguments{
\item{x}{a \code{linelist} object}

\item{allow_extra}{a \code{logical} indicating if additional data types not
currently recognized by \code{linelist} should be allowed; if so, corresponding
tags will be added}
}
\value{
If checks pass, a \code{linelist} object; otherwise issues an error.
}
\description{
This function evalutes the validity of the tags of a \code{linelist} object by
checking that: i) tags are present ii) tags is a \code{list} of \code{character} iii)
that all default tags are present iv) tagged variables exist v) that no extra
tag exists (if \code{allow_extra} is \code{FALSE}).
}
\seealso{
\url{validate_types} to check if tagged variables have
the right classes
}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
