% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_type.R
\name{validate_type}
\alias{validate_type}
\title{Check a tagged variable has the right type}
\usage{
validate_type()
}
\value{
A named \code{list}.
}
\description{
This function checks that a specific tagged variable has the expected type of
data. Default expected types are defined by \url{tags_types}.
}
\examples{

}
\seealso{
\url{tags_types} to check or change acceptable data
classes for specific tags
}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
