% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_types.R
\name{validate_types}
\alias{validate_types}
\title{Type check variables}
\usage{
validate_types(x, ref_types = vars_types(), strict = FALSE)
}
\arguments{
\item{x}{a \code{linelist} object}

\item{ref_types}{a named \code{list} providing variable names and their allowed
types. Defaults to output from by \code{\link[=vars_types]{vars_types()}}, which can be used to update
naming of default variables, for example \code{vars_types(id = "case_ID")}. Extra
variables can be concatenated, for example
\code{c(vars_types(), y_loc = type("numeric"))}.}

\item{strict}{a \code{logical} indicating whether all defaults must be present
(\code{TRUE}) or not (\code{FALSE})}
}
\value{
A named \code{list}.
}
\description{
This function checks the class of variables in a \code{linelist} against
pre-defined accepted classes in \code{\link[=vars_types]{vars_types()}} and further specifications.
}
\examples{
if (require(outbreaks) && require(magrittr)) {
  x <- make_linelist(
    measles_hagelloch_1861,
    !!!update_defaults(id = "case_ID")
  )
  validate_types(x, ref_types = c(vars_types(), y_loc = type("numeric")))

  ## create an invalid linelist - onset date is a factor
  x <- measles_hagelloch_1861 \%>\%
    make_linelist(!!!update_defaults(gender = "date_of_prodrome"))
  x

  ## the below issues an error
  ## note: tryCatch is only used to avoid a genuine error in the example
  tryCatch(validate_types(x,
    ref_types = vars_types(gender = "date_of_prodrome")
  ), error = paste)
}
}
\seealso{
\itemize{
\item \code{\link[=type]{type()}} for classes of types
\item \code{\link[=vars_types]{vars_types()}} to change allowed types
\item \code{\link[=validate_labels]{validate_labels()}} to perform a series of checks on the tags
\item \code{\link[=validate_linelist]{validate_linelist()}} to combine \code{validate_labels} and \code{validate_types}
}
}
