% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_types.R
\name{vars_types}
\alias{vars_types}
\title{Generate acceptable variable types for labels}
\usage{
vars_types(...)
}
\arguments{
\item{...}{A series of column names provided as \verb{col1, col2} to add columns
or optionally \verb{id = 'ID', date_onset = 'date_symptom_onset'} to rename
defaults}
}
\value{
A named \code{list} of variable names and type values.
}
\description{
This function returns a named list providing the acceptable data types for
the default labels. If no argument is provided, it returns default
values. Otherwise, provided values will be used to define the defaults.
}
\examples{
# list default values
vars_types()

# rename default variable
vars_types(id = "case_ID")

}
\seealso{
\itemize{
\item \code{\link[=label_defaults]{label_defaults()}} for the default tags
\item \code{\link[=validate_labels]{validate_labels()}} uses \code{\link[=vars_types]{vars_types()}} for validating tags
\item \code{\link[=validate_linelist]{validate_linelist()}} uses \code{\link[=vars_types]{vars_types()}} for validating tags
}
}
