% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vars.R, R/set_vars.R
\name{list_vars}
\alias{list_vars}
\alias{get_vars}
\alias{date_of_onset}
\alias{date_onset}
\alias{id}
\alias{gender}
\alias{date_report}
\alias{age}
\alias{age_group}
\alias{geo}
\alias{set_vars}
\alias{set_vars<-}
\alias{set_vars.linelist}
\alias{set_vars<-.linelist}
\title{Accessor functfons}
\usage{
list_vars(x)

get_vars(x, ..., simplify = TRUE)

date_of_onset(x)

date_onset(x)

id(x)

gender(x)

date_report(x)

age(x)

age_group(x)

geo(x)

set_vars(x, ...)

set_vars(x, name) <- value

\method{set_vars}{linelist}(x, ...)

\method{set_vars}{linelist}(x, name) <- value
}
\arguments{
\item{x}{a linelist object}

\item{...}{For \code{set_vars()}, any number of variables defined in
\code{\link[=epivars]{epivars()}} that can be used for mapping or modelling.}

\item{simplify}{if \code{TRUE} (default) and if the result is a single column,
then this variable will be returned as a \code{vector}; otherwise (\code{FALSE}), a
\code{data.frame} is returned.}

\item{name}{the name of the variable in \code{\link[=epivars]{epivars()}} to assign}

\item{value}{the name of the column in the locations data}

\item{var}{the name of a variable as a string}
}
\value{
a vector if the epivar can be represented by a single column, a data
frame otherwise (e.g. a location column).
}
\description{
This set of fucntions defines helper functions for accessing pre-defined
epivars in your linelist object.
}
\examples{

toy_data <- messy_data()
cleaned_data <- clean_data(toy_data)
ll <- as_linelist(cleaned_data, 
                  id = "id", 
                  date_onset = "date_of_onset",
                  gender = "gender",
                  geo = c("lon", "lat"))
date_of_onset(ll)
id(ll)
gender(ll)
geo(ll)
get_var(ll, "geo", vector = FALSE)
# epivars that haven't been defined for the data set will return an error
try(date_report(ll))
}
