% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_epivars.R
\name{get_epivars}
\alias{get_epivars}
\alias{id}
\alias{date_onset}
\alias{date_report}
\alias{date_outcome}
\alias{case_definition}
\alias{outcome}
\alias{gender}
\alias{age}
\alias{age_group}
\alias{geo}
\title{Accessor functfons}
\usage{
get_epivars(x, ..., simplify = TRUE)

id(x)

date_onset(x)

date_report(x)

date_outcome(x)

case_definition(x)

outcome(x)

gender(x)

age(x)

age_group(x)

geo(x)
}
\arguments{
\item{x}{a linelist object}

\item{...}{\emph{epivars} to be used; can be provided as characters, integers,
or logical; integers and logicals are to be matched against the vector of
epi variables returned by \code{list_epivars(x, simple = TRUE)}. If nothing is
provided, all \emph{epivars} are returned.}

\item{simplify}{if \code{TRUE} (default) and a single epivar is specified, it will
be converted to a \code{vector}. Otherwise, a data frame is returned.}
}
\value{
a vector if the epivar can be represented by a single column, a data
frame otherwise (e.g. a location column).
}
\description{
This set of fucntions defines helper functions for accessing pre-defined
epivars in your linelist object.
}
\note{
For the "geo" epivar, which necessarily contains two columns for
longitude and latitude data, when \code{simplify = TRUE}, a matrix will be
returned.
}
\examples{

## make toy data
toy_data <- messy_data()
cleaned_data <- clean_data(toy_data)
ll <- as_linelist(cleaned_data, 
                  id = "id", 
                  date_onset = "date_of_onset",
                  gender = "gender",
                  geo = c("lon", "lat"))

## general purpose accessor
list_epivars(ll, simple = TRUE)
get_epivars(ll) # all epi variable
get_epivars(ll, "id", "date_onset", "gender") # named epi variables
date_of_onset(ll)
id(ll)
gender(ll)
geo(ll) # a matrix with two columns
get_epivars(ll, "geo", simplify = FALSE) # a data frame with two columns

# epivars that haven't been defined for the data set will return an error
try(date_report(ll))
}
