% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linelist_class.R
\name{as_linelist}
\alias{as_linelist}
\alias{as_linelist.default}
\alias{as_linelist.data.frame}
\title{Create a linelist object}
\usage{
as_linelist(dat, ...)

\method{as_linelist}{default}(dat, ...)

\method{as_linelist}{data.frame}(dat, ...)
}
\arguments{
\item{dat}{a data frame}

\item{...}{options passed to \code{\link[=set_vars]{set_vars()}}}
}
\description{
Create a linelist object
}
\examples{
md <- messy_data()
cd <- clean_data(md)
ll <- as_linelist(cd, 
                  id = "id", 
                  date_onset = "date_of_onset", 
                  gender = "gender",
                  geo = c("lon", "lat")
                 )
ll
class(ll)
}
