% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_variable_labels.R
\name{clean_variable_labels}
\alias{clean_variable_labels}
\title{Clean variable labels}
\usage{
clean_variable_labels(x, classes = NULL, ...)
}
\arguments{
\item{x}{a \code{data.frame}}

\item{classes}{a vector of class definitions for each of the columns. If this
is not provided, the classes will be read from the columns themselves.
Practically, this is used in \code{\link[=clean_data]{clean_data()}} to mark columns as protected.}

\item{...}{further arguments passed to \code{\link[epitrix:clean_labels]{epitrix::clean_labels()}}; the most
important is \code{sep}, which refers to the separator used between words,
and defaults to the underscore \code{_}.}
}
\value{
A \code{data.frame} with standardised labels for characters and
factors.
}
\description{
This function standardises the labels of all characters of factors in a
\code{data.frame}. It uses the standardisation implemented by
\code{\link[epitrix:clean_labels]{epitrix::clean_labels()}} in the \code{epitrix} package. See
\code{?epitrix::clean_labels} for more information.
}
\examples{

## make toy data
onsets <- as.Date("2018-01-01") + sample(1:10, 20, replace = TRUE)
genders <- c("male", "female", "FEMALE", "Male", "Female", "MALE")
gender <- sample(genders, 20, replace = TRUE)
case_types <- c("confirmed", "probable", "suspected", "not a case",
                "Confirmed", "PROBABLE", "suspected  ", "Not.a.Case")
case <- factor(sample(case_types, 20, replace = TRUE))
toy_data <- data.frame("Date of Onset." = onsets,
                       "GENDER_ " = gender,
                       "Épi.Case_définition" = case)
## show data
toy_data


## clean variable names, store in new object, show results
clean_data <- clean_variable_names(toy_data)
clean_data
}
\author{
Thibaut Jombart
}
