% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_dictionary.R, R/dictionary.R
\name{default_dictionary}
\alias{default_dictionary}
\alias{linelist_options}
\alias{get_dictionary}
\alias{set_dictionary}
\alias{reset_dictionary}
\title{Linelist Global Variables}
\usage{
default_dictionary()

linelist_options(..., set = FALSE, reset = FALSE)

get_dictionary(...)

set_dictionary(..., reset = FALSE)

reset_dictionary(...)
}
\arguments{
\item{...}{quoted varaibles to add to the default variables}

\item{set}{when \code{TRUE}, the variables provided in \code{...} will be added to the
global variables. Defaults to `FALSE``}

\item{reset}{when \code{TRUE}, the global variables are reset to the default
variables listed above. Defaults to \code{FALSE}}
}
\description{
This defines the legal variables that linelist will recognise.
}
\examples{

# see the default varaibles
get_dictionary()

# Equivalent
getOption("linelist_epivars")

# If you create a new method and need other varaibles, or just want a shorter
# representation, they can be added to your options:

get_dictionary("random", set = TRUE)

# You can also reset the variables
get_dictionary(reset = TRUE)
}
