% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{default_dictionary}
\alias{default_dictionary}
\alias{default_dictionary_path}
\alias{get_dictionary}
\alias{set_dictionary}
\alias{reset_dictionary}
\title{Linelist Global Variables}
\usage{
default_dictionary()

default_dictionary_path()

get_dictionary()

set_dictionary(x, ...)

reset_dictionary()
}
\arguments{
\item{x}{either a data frame or the path to a that can produce a data frame.}

\item{...}{parameters passed on to \code{\link[rio:import]{rio::import()}}.}
}
\description{
These functions let the user define variables that are important
to a given analysis script. These functions add to or reset variables
contained in \code{getOption("linelist_epivars")}.
\itemize{
\item \code{set_dictionary()}  will \strong{add} variables to the current defined epivars.
\strong{Use this function at the beginning of your analysis script or Rmd file.}
\item \code{get_dictionary()} shows the current defined epivars that linelist will
recognise.
\item \code{reset_dictionary()} will reset the dictionary to those in
\code{default_dictionary()}.
}

All of these functions are wrappers for \code{linelist_options()}, which itself
is a wrapper for \code{options(linelist_epivars = ...)}.
}
\details{
This defines the legal variables that linelist will recognise.

The expected format for these data is a\code{data.frame} with 3 columns
'epivar', 'hxl' and 'description', representing the new dictionary to be
used; A template example can be found by using \code{\link[=default_dictionary]{default_dictionary()}}.
}
\examples{
# see the default varaibles
get_dictionary()

# Equivalent
getOption("linelist_dictionary")


# Set a new, one-column dictionary 
hosp <- data.frame(
  epivar = "date_hospital",
  hxl    = "#date +start",
  description = "date at which patient was hospitalized",
  stringsAsFactors = FALSE
)

set_dictionary(hosp)
get_dictionary()

# Use the default_dictionary_path() to read in the default dictionary from
# disk
default_dictionary_path()
set_dictionary(default_dictionary_path())
get_dictionary()

# You can also reset the variables automatically
reset_dictionary()

}
\author{
Zhian N. Kamvar, Thibaut Jombart
}
