% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{default_dictionary}
\alias{default_dictionary}
\alias{get_dictionary}
\alias{set_dictionary}
\alias{reset_dictionary}
\title{Linelist Global Variables}
\usage{
default_dictionary()

get_dictionary()

set_dictionary(x)

reset_dictionary()
}
\arguments{
\item{x}{a \code{data.frame} with 3 columns 'epivar', 'hxl' and 'description',
representing the new dictionary to be used; see \code{default_dictionary} for a
template}

\item{...}{quoted varaibles to add to the current dictionary}

\item{set}{when \code{TRUE}, the variables provided in \code{...} will be added to the
global variables. Defaults to `FALSE``}

\item{reset}{when \code{TRUE}, the global variables are reset to the default
variables listed above. Defaults to \code{FALSE}}
}
\description{
These functions let the user define variables that are important
to a given analysis script. These functions add to or reset variables
contained in \code{getOption("linelist_epivars")}.
\itemize{
\item \code{set_dictionary()}  will \strong{add} variables to the current defined epivars.
\strong{Use this function at the beginning of your analysis script or Rmd file.}
\item \code{get_dictionary()} shows the current defined epivars that linelist will
recognise.
\item \code{reset_dictionary()} will reset the dictionary to those in
\code{default_dictionary()}.
}

All of these functions are wrappers for \code{linelist_options()}, which itself
is a wrapper for \code{options(linelist_epivars = ...)}.
}
\details{
This defines the legal variables that linelist will recognise.
}
\examples{
# see the default varaibles
get_dictionary()

# Equivalent
getOption("linelist_epivars")


# You can also reset the variables
reset_dictionary()

}
\author{
Zhian N. Kamvar, Thibaut Jombart
}
