% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_dates.R
\name{guess_dates}
\alias{guess_dates}
\title{Try and guess dates from a characters}
\usage{
guess_dates(x, error_tolerance = 0.1, first_date = NULL,
  last_date = Sys.Date(), orders = list(world = c("dby", "dmy", "Ybd",
  "Ymd"), US = c("Omdy", "YOmd")), quiet = TRUE)
}
\arguments{
\item{x}{a \code{character} vector or a \code{factor}}

\item{error_tolerance}{a number between 0 and 1 indicating the proportion of
entries which cannot be identified as dates to be tolerated; if this
proportion is exceeded, the original vector is returned, and a message is
issued; defaults to 0.1 (10 percent)}

\item{first_date}{a Date object specifying the first valid date. Defaults to
fifty years before the \code{last_date}.}

\item{last_date}{a Date object specifying the last valid date. Defaults to the
current date.}

\item{orders}{date codes for fine-grained parsing of dates. This allows for
parsing of mixed dates. If a list is supplied, that list will be used for
successive tries in parsing.  This is passed on to
\code{\link[lubridate:parse_date_time]{lubridate::parse_date_time()}}. Default orders parse World dmy/dby dates
before US mdy/bdy dates.}

\item{quiet}{a logical indicating if messages should be displayed to the
console (\code{TRUE}, default); set to \code{FALSE} to silence messages}
}
\description{
Note that THIS FEATURE IS STILL EXPERIMENTAL: we strongly recommend checking
a few converted dates manually. This function tries to extract dates from a
\code{character} vector or a \code{factor}. It treats each entry independently, using
regular expressions to detect if a date is present, its format, and if
successful it converts that entry to a standard \code{Date} with the \emph{Ymd} format
(e.g. \code{2018-01-21}). Entries which cannot be processed result in \code{NA}. An
error threshold can be used to define the maximum number of resulting \code{NA}
(i.e. entries without an identified date) that can be tolerated. If this
threshold is exceeded, the original vector is returned.
}
\details{
Converting ambiguous character strings to dates is difficult for
many reasons:
\itemize{
\item dates may not use the standard Ymd format
\item within the same variable, dates may follow different formats
\item dates may be mixed with things that are not dates
\item the behaviour of \code{as.Date} in the presence of non-date is hard to predict,
sometimes returning \code{NA}, sometimes issuing an error.
}

This function tries to address all the above issues. Dates with the following
format should be automatically detected, irrespective of separators
(e.g. "-", " ", "/") and surrounding text:
\itemize{
\item "19 09 2018"
\item "2018 09 19"
\item "19 Sep 2018"
\item "2018 Sep 19"
}

Note that if a character string has multiple dates, it is currently hard to
predict which date will be returned.
}
\examples{

x <- c("01-12-2001", "male", "female", "2018-10-18", NA, NA, "2018_10_17",
      "2018 10 19", "// 24/12/1989", "this is 24/12/1989!",
      "RECON NGO: 19 Sep 2018 :)", "6/9/11", "10/10/10")
FIRST_DATE <- as.Date("1969-11-11")
guess_dates(x, error_tolerance = 1, first_date = FIRST_DATE) # forced conversion
guess_dates(x, error_tolerance = 0.15, first_date = FIRST_DATE) # only 15\% errors allowed
}
\author{
Thibaut Jombart
}
