% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_defaults.R
\name{label_defaults}
\alias{label_defaults}
\title{Generate default tags for a linelist}
\usage{
label_defaults()
}
\value{
A named \code{list}.
}
\description{
This function returns a named list providing the default tags for a
\code{linelist} object (all default to NULL).
}
\examples{
label_defaults()

}
