% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linelist-package.R
\docType{package}
\name{linelist-package}
\alias{linelist}
\alias{linelist-package}
\title{linelist: Tools to Import and Tidy Case Linelist Data}
\description{
A collection of wrappers for importing case linelist data from usual formats, and tools for cleaning data, detecting dates, and storing meta-information on the content of the resulting data frame.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/reconhub/linelist}
  \item Report bugs at \url{https://github.com/reconhub/linelist/issues}
}

}
\author{
\strong{Maintainer}: Thibaut Jombart \email{thibautjombart@gmail.com}

Authors:
\itemize{
  \item Zhian N. Kamvar \email{zkamvar@gmail.com}
}

Other contributors:
\itemize{
  \item Patrick Barks \email{patrick.barks@epicentre.msf.org} [contributor]
}

}
\keyword{internal}
