% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_epivars.R
\name{list_epivars}
\alias{list_epivars}
\title{Retrieve metadata from a linelist object}
\usage{
list_epivars(x, simple = FALSE, epivars_only = FALSE,
  dictionary = TRUE)
}
\arguments{
\item{x}{a \code{linelist} object}

\item{simple}{a logical indicating if a single vector of names of present
\code{epivars} should be returned (\code{TRUE}); otherwise the output will contain
more information as a \code{data.frame} (\code{FALSE}, default)}

\item{epivars_only}{a \code{logical} indicating if columns which are not registers
\code{epivars} should be removed from the output}

\item{dictionary}{if \code{TRUE}, a column called "\emph{epivar}" is appended to the
metadata table, indicating which standard epi variable columns correspond
to; of \code{FALSE}, the class of the column replaces it}
}
\value{
if \code{simple} is \code{TRUE}, a vector of characters giving the names of
available \code{epivariables}; otherwise, a \code{data.frame} giving more information
about these variables
}
\description{
Retrieve metadata from a linelist object
}
\examples{
dat <- clean_data(messy_data())
ll  <- as_linelist(dat,
                  id = "id", 
                  date_onset = "date_of_onset", 
                  gender = "gender",
                  geo = c("lon", "lat")
                 )
list_epivars(ll)
list_epivars(ll, TRUE)
list_epivars(ll, epivars_only = TRUE)
list_epivars(ll, dictionary = FALSE)
}
