% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_linelist.R
\name{make_linelist}
\alias{make_linelist}
\title{Title of the function}
\usage{
make_linelist(x)
}
\arguments{
\item{x}{a \code{data.frame} or a \code{tibble} containing case line list data, with
cases in rows and variables in columns}
}
\value{
The function returns a \code{linelist} object.
}
\description{
A short description of what it does, the main inputs and outputs.
}
\details{
Some more information on the function
}
\examples{
# basic use of the function
template(1)

}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
