% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_linelist.R
\name{make_linelist}
\alias{make_linelist}
\title{Create a linelist from a data.frame}
\usage{
make_linelist(x)
}
\arguments{
\item{x}{a \code{data.frame} or a \code{tibble} containing case line list data, with
cases in rows and variables in columns}
}
\value{
The function returns a \code{linelist} object.
}
\description{
This function converts a \code{data.frame} or a \code{tibble} to a \code{linelist} object,
where different types of epidemiologically relevant data are tagged. This
includes dates of different events (e.g. onset of symtpom, case reporting),
information on the patient (e.g. age, gender, location) as well as other
informations such as the type of case (e.g. confirmed, probable) or the
outcome of the disease. The output will seem to be the same \code{data.frame}, but
\code{linelist}-aware packages will then be able to automatically use tagged
fields for further data cleaning and analysis.
}
\details{
Some more information on the function
}
\examples{
# basic use of the function
template(1)

}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
