% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.linelist.R
\name{print.linelist}
\alias{print.linelist}
\title{Printing method for linelist objects}
\usage{
\method{print}{linelist}(x, ...)
}
\arguments{
\item{x}{a \code{linelist} object}

\item{...}{further arguments to be passed to 'print'}
}
\value{
Invisibly returns the object.
}
\description{
This function prints linelist objects
}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
