% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_tag.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{has_tag}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{safeframe}{\code{\link[safeframe]{has_tag}}}
}}

