% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.linelist}
\alias{select.linelist}
\alias{select_id}
\alias{select_date_onset}
\alias{select_date_reporting}
\alias{select_date_admission}
\alias{select_date_discharge}
\alias{select_date_outcome}
\alias{select_date_death}
\alias{select_gender}
\alias{select_age}
\alias{select_location}
\alias{select_occupation}
\alias{select_hcw}
\alias{select_outcome}
\title{Subset columns of a linelist object}
\usage{
\method{select}{linelist}(.data, ..., tags = NULL, lost_action = "warning")

select_id(x)

select_date_onset(x)

select_date_reporting(x)

select_date_admission(x)

select_date_discharge(x)

select_date_outcome(x)

select_date_death(x)

select_gender(x)

select_age(x)

select_location(x)

select_occupation(x)

select_hcw(x)

select_outcome(x)
}
\arguments{
\item{.data}{a \code{linelist} object}

\item{...}{the variables to select, either using their column names, or tag
names (or a mixture)}

\item{tags}{a \code{character} indicating tagged variables to select using tag
names (see \code{tags_names()}) for default values}

\item{lost_action}{a \code{character} indicating the behaviour to adopt when some
of the tagged variables are dropped through the \code{select} process; can be
"warning" (default), "error", or "none".}

\item{x}{a \code{linelist} object}
}
\value{
The function returns a \code{linelist} with selected columns.
}
\description{
This function works similarly to \code{dplyr::select} but can in addition refer to
tagged variables through the \code{tags} argument. When variables are selected
using both procedures, tagged variables are output as the last columns.
}
\seealso{
\itemize{
\item \url{select_tags} to select tags only
\item \url{tags_df} to return a \code{data.frame} of all tagged variables
}
}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
