% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.linelist}
\alias{select.linelist}
\title{subset columns of a linelist object}
\usage{
\method{select}{linelist}(.data, ..., warn = TRUE)
}
\arguments{
\item{.data}{a \code{linelist} object}

\item{...}{the variables to select, either using their column names, or tag
names (or a mixture)}

\item{warn}{a \code{logical} indicating if a warning should be issued if some
tagged variables have been lost by the \code{select} operation; defaults to
\code{TRUE}}
}
\value{
The function returns a \code{linelist} with selected columns.
}
\description{
This function works similarly to \code{dplyr::select} but can in addition refer to
tagged variables.
}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
