% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.linelist}
\alias{select.linelist}
\title{subset columns of a linelist object}
\usage{
\method{select}{linelist}(.data, ...)
}
\arguments{
\item{x}{a \code{linelist} object}
}
\value{
The function returns a \code{linelist} with selected columns.
}
\description{
This function works similarly to \code{dplyr::select} but can in addition refer to
tagged variables.
}
\examples{
# basic use of the function
if (require(outbreaks)) {
measles_hagelloch_1861
x <- make_linelist(measles_hagelloch_1861, date_onset = "date_of_prodrome")
tags(x)
}

}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
