% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.linelist}
\alias{select.linelist}
\title{subset columns of a linelist object}
\usage{
\method{select}{linelist}(.data, ..., warn = TRUE)
}
\arguments{
\item{.data}{a \code{linelist} object}

\item{...}{the variables to select, either using their column names, or tag
names (or a mixture)}

\item{warn}{a \code{logical} indicating if a warning should be issued if some
tagged variables have been lost by the \code{select} operation; defaults to
\code{TRUE}}
}
\value{
The function returns a \code{linelist} with selected columns.
}
\description{
This function works similarly to \code{dplyr::select} but can in addition refer to
tagged variables. We recommend referring to variables and tags by their
names: using \code{logical} and \code{integer} may have unexecpted results due to how
regular variables and tagged variables are combined (see details).
}
\details{
The function relies on a concatenation of two data.frames, the first
one containing the regular variables, and the second one containing tagged
variables. This approach is safe as long as regular variables and tagged
variables are referred to by their names, which is the recommended
approach. Other approaches using \code{integer} or \code{logical} may yield more
confusing outputs, e.g. selecting \code{all} variables using \code{TRUE} will return
the two data.frames concatenated.
}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
