% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.linelist}
\alias{select.linelist}
\title{Subset columns of a linelist object}
\usage{
\method{select}{linelist}(.data, ..., tags = NULL, lost_action = "warning")
}
\arguments{
\item{.data}{a \code{linelist} object}

\item{...}{the variables to select, using \code{dplyr} compatible syntax}

\item{tags}{a \code{character} indicating tagged variables to select using tag
names (see \code{tags_names()}) for default values; values can be named, in
which case the output columns will be renamed accordingly (e.g. \code{onset = "date_onset"} will output a column named 'onset').}

\item{lost_action}{a \code{character} indicating the behaviour to adopt when some
of the tagged variables are dropped through the \code{select} process; can be
"warning" (default), "error", or "none".}
}
\value{
The function returns a \code{linelist} with selected columns.
}
\description{
This function works similarly to \code{dplyr::select} but can in addition refer to
tagged variables through the \code{tags} argument. When variables are selected
using both procedures, tagged variables are output as the last columns.
}
\seealso{
\itemize{
\item \url{select_tags} to select tags only
\item \url{tags_df} to return a \code{data.frame} of all tagged variables
}
}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
