% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_tags.R
\name{select_tags}
\alias{select_tags}
\title{Extract tagged variables of a linelist object}
\usage{
select_tags(x, ...)
}
\arguments{
\item{x}{a \code{linelist} object}

\item{...}{the tagged variables to select, using \code{dplyr::select} compatible
terminology; see \url{tags_names}) for default values}
}
\value{
A \code{data.frame} of tagged variables.
}
\description{
This function is used to create a \code{data.frame} of tagged variables from a
\code{linelist} object, and runs \code{dplyr::select} on the output. It is equivalent
to running successively \code{tags_df} and \code{dplyr::select} on a \code{linelist}
object. Note that the output no longer is a \code{linelist} object, but a regular
\code{data.frame}.
}
\seealso{
\url{tags_names}) for the names of tags
}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
