% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_epivars.R
\name{set_epivars}
\alias{set_epivars}
\alias{add_epivars}
\title{Define which columns are epivars in a linelist}
\usage{
set_epivars(x, ...)

add_epivars(x, ...)
}
\arguments{
\item{x}{a \code{linelist} object}

\item{...}{a series of column names named after epivars}
}
\description{
These functions can be used to change the \code{epivars} registred in a \code{linelist}
object. \code{set_epivars} replaces previous \code{epivars} with new
values. \code{add_epivars} adds new \code{epivars} to the existing ones.
}
