% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_linelist.R
\name{validate_linelist}
\alias{validate_linelist}
\title{Checks the content of a linelist object}
\usage{
validate_linelist(
  x,
  strict = FALSE,
  allow_extra = TRUE,
  ref_types = vars_types()
)
}
\arguments{
\item{x}{a \code{linelist} object}

\item{strict}{a \code{logical} indicating whether all defaults must be present
(\code{TRUE}) or not (\code{FALSE})}

\item{allow_extra}{a \code{logical} indicating whether non-default variables are
permitted}

\item{ref_types}{a named \code{list} providing variable names and their allowed
types. Defaults to output from by \code{\link[=vars_types]{vars_types()}}, which can be used to update
naming of default variables, for example \code{vars_types(id = "case_ID")}. Extra
variables can be concatenated, for example
\code{c(vars_types(), y_loc = type("numeric"))}.}
}
\value{
If checks pass, a \code{linelist} object (invisibly); otherwise issues an
error.
}
\description{
This function evaluates the validity of a \code{linelist} object by checking the
object class, its labels, and the types of the labelled variables. It
combines validations checks made by \code{\link[=validate_types]{validate_types()}} and
\code{\link[=validate_labels]{validate_labels()}}. See 'Details' section for more information on the
checks performed.
}
\details{
The following checks are performed:
\itemize{
\item \code{x} is a \code{linelist} object
\item \code{x} has well-formed labels
\item all default labels are present (even if \code{NULL})
\item all labelled variables correspond to existing columns
\item all labelled variables have an acceptable class
\item (optional) \code{x} has no extra label beyond the default labels
}

Please note that if you deviate from default variable naming, you will also
need to update them while validating the linelist.
}
\examples{

if (require(outbreaks) && require(magrittr)) {
  ## create a valid linelist
  x <- measles_hagelloch_1861 \%>\%
    make_linelist()
  x

  ## rename the id default
  x <- make_linelist(
    measles_hagelloch_1861,
    !!!update_defaults(id = "case_ID")
  )

  ## validation
  validate_linelist(x)

  ## create an invalid linelist - onset date is a factor
  x <- measles_hagelloch_1861 \%>\%
    make_linelist(!!!update_defaults(gender = "date_of_prodrome"))
  x

  ## the below issues an error
  ## note: tryCatch is only used to avoid a genuine error in the example
  tryCatch(validate_linelist(x,
    ref_types = vars_types(gender = "date_of_prodrome")
  ), error = paste)
}
}
\seealso{
\itemize{
\item \code{\link[=vars_types]{vars_types()}} to change allowed types
\item \code{\link[=validate_types]{validate_types()}} to check if labelled variables have the right classes
\item \code{\link[=validate_labels]{validate_labels()}} to perform a series of checks on the labels
}
}
