% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_linelist.R
\name{validate_linelist}
\alias{validate_linelist}
\title{Checks the content of a linelist object}
\usage{
validate_linelist(x, ref_types = tags_types(), allow_extra = FALSE)
}
\arguments{
\item{x}{a \code{linelist} object}

\item{ref_types}{a \code{list} providing allowed types for all tags, as returned
by \url{tags_types}}

\item{allow_extra}{a \code{logical} indicating if additional data types not
currently recognized by \code{linelist} should be allowed; if so, corresponding
tags will be added}
}
\value{
If checks pass, a \code{linelist} object; otherwise issues an error.
}
\description{
This function evalutes the validity of a \code{linelist} object by checking the
object class, its tags, and the types of the tagged variables.
}
\seealso{
\url{validate_types} to check if tagged variables have
the right classes
}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
