% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_types.R
\name{validate_types}
\alias{validate_types}
\title{Check tagged variables are the right class}
\usage{
validate_types(x, ref_types = tags_types())
}
\arguments{
\item{x}{a \code{linelist} object}

\item{ref_types}{a \code{list} providing allowed types for all tags, as returned
by \url{tags_types}}
}
\value{
A named \code{list}.
}
\description{
This function checks the class of each tagged variable in a \code{linelist}
against pre-defined accepted classes in \href{tags_types}{\code{tags_types()}}.
}
\examples{
if (require(outbreaks) && require(dplyr) && require(magrittr)) {

  ## create an invalid linelist - gender is a numeric
  x <- measles_hagelloch_1861 \%>\%
    tibble() \%>\% 
    make_linelist(id = "case_ID",
                  gender = "infector")
  x

  ## the below would issue an error
  # validate_types(x)

  ## to allow other types, e.g. gender to be integer, character or factor
  validate_types(x, tags_types(gender = c("integer", "character", "factor")))
}
}
\seealso{
\url{tags_types} to check or change acceptable data
classes for specific tags; \url{validate_type} to apply
validation to a single variable (non exported)
}
\author{
Thibaut Jombart \url{thibaut@data.org}
}
